refactor(ccxt): 統一 CCXT 工廠，刪除 CcxtInstanceManager
- 刪除 CcxtInstanceManager，功能整合至 CcxtExchangeFactory
- 新增靜態方法 createPublicExchangeWithCache() 供公開查詢使用
- 新增靜態方法 createAuthenticatedExchangeForQuery() 供認證查詢使用
- PositionDetailsService 改用 CcxtExchangeFactory 靜態方法
- 保留全局 markets 快取與帳戶類型快取機制
